Separating the JavaFX related classes into their own subpackage
