+++ platform/api.htmlui/src/org/netbeans/modules/htmlui/HTMLDialogImpl.java
@@ -22,24 +22,11 @@ import java.awt.Dialog;
 import java.awt.Dimension;
 import java.awt.EventQueue;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javafx.application.Platform;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
-import javafx.embed.swing.JFXPanel;
-import javafx.scene.Node;
-import javafx.scene.Scene;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.paint.Color;
-import javafx.scene.web.WebView;
-import javafx.stage.Stage;
 import javax.swing.JButton;
 import javax.swing.JComponent;
 import org.openide.*;
@@ -47,12 +34,10 @@ import org.openide.util.Exceptions;
 import org.openide.util.Lookup;
 
 public final class HTMLDialogImpl implements Runnable {
-    private static final Logger LOG = Logger.getLogger(HTMLDialogImpl.class.getName());
-    
     private volatile int state;
-    private NbFxPanel p;
+    private JComponent p;
     private DialogDescriptor dd;
-    private WebView webView;
+    private Object webView;
     
     private String url;
     private Runnable onPageLoad;
@@ -78,7 +63,8 @@ public final class HTMLDialogImpl implements Runnable {
                 initPanel();
                 break;
             case 1:
-                initFX();
+                state = 2;
+                webView = HtmlToolkit.getDefault().initHtmlDialog(url, dd, p, this, techIds);
                 break;
             case 2:
                 initPage();
@@ -92,11 +78,11 @@ public final class HTMLDialogImpl implements Runnable {
             case 3:
                 showDialog();
                 state = 4;
-                Platform.runLater(this);
+                HtmlToolkit.getDefault().execute(this);
                 break;
             case 4:
                 state = -1;
-                exitNestedLoop(this);
+                HtmlToolkit.getDefault().exitNestedLoop(this);
                 break;
             default:
                 throw new IllegalStateException("State: " + state);
@@ -108,10 +94,10 @@ public final class HTMLDialogImpl implements Runnable {
             run();
             showDialog();
         } else {
-            if (Platform.isFxApplicationThread()) {
+            if (HtmlToolkit.getDefault().isApplicationThread()) {
                 nestedLoop = true;
                 EventQueue.invokeLater(this);
-                enterNestedLoop(this);
+                HtmlToolkit.getDefault().enterNestedLoop(this);
             } else {
                 try {
                     EventQueue.invokeAndWait(this);
@@ -132,60 +118,13 @@ public final class HTMLDialogImpl implements Runnable {
     }
     
     private void initPanel() {
-        p = new NbFxPanel();
+        p = HtmlToolkit.getDefault().newPanel();
         dd = new DialogDescriptor(p, "");
         dd.setOptions(new Object[0]);
         state = 1;
-        Platform.runLater(this);
+        HtmlToolkit.getDefault().execute(this);
     }
     
-    private void initFX() {
-        Platform.setImplicitExit(false);
-        webView = new WebView();
-        BorderPane bp = new BorderPane();
-        Scene scene = new Scene(bp, Color.ALICEBLUE);
-
-        class X implements ChangeListener<String>, Runnable {
-
-            private String title;
-
-            public X() {
-                super();
-            }
-
-            @Override
-            public void changed(ObservableValue<? extends String> ov, String t, String t1) {
-                title = webView.getEngine().getTitle();
-                EventQueue.invokeLater(this);
-            }
-
-            @Override
-            public void run() {
-                if (title != null) {
-                    dd.setTitle(title);
-                }
-            }
-        }
-        final X x = new X();
-        webView.getEngine().titleProperty().addListener(x);
-        Platform.runLater(x);
-        bp.setCenter(webView);
-        p.setScene(scene);
-
-        ClassLoader loader = Lookup.getDefault().lookup(ClassLoader.class);
-        if (loader == null) {
-            loader = HTMLDialogImpl.class.getClassLoader();
-        }
-        URL pageUrl;
-        try {
-            pageUrl = new URL(url);
-        } catch (MalformedURLException ex) {
-            throw new IllegalStateException(ex);
-        }
-        state = 2;
-        NbBrowsers.load(webView, pageUrl, this, loader, techIds.toArray());
-    }
-
     private void initPage() {
         try {
             onPageLoad.run();
@@ -208,72 +147,6 @@ public final class HTMLDialogImpl implements Runnable {
         } catch (MalformedURLException ex) {
             throw new IllegalStateException(ex);
         }
-        if (type == Node.class) {
-            WebView wv = new WebView();
-            NbBrowsers.load(wv, pageUrl, onPageLoad, loader, techIds.toArray());
-            return type.cast(wv);
-        } else if (type == JComponent.class) {
-            final JFXPanel tmp = new JFXPanel();
-            final ClassLoader l = loader;
-            Platform.runLater(new Runnable() {
-                @Override
-                public void run() {
-                    WebView wv = new WebView();
-                    NbBrowsers.load(wv, pageUrl, onPageLoad, l, techIds.toArray());
-                    Scene s = new Scene(wv);
-                    tmp.setScene(s);
-                }
-            });
-            return type.cast(tmp);
-        } else {
-            throw new IllegalStateException("Unsupported type: " + type);
-        }
-    }
-
-    private static final Method GET;
-    private static final Method ENTER;
-    private static final Method EXIT;
-    static {
-        Method g = null;
-        Method n = null;
-        Method x = null;
-        try {
-            Class<?> tC;
-            final String toolkitCN = "com.sun.javafx.tk.Toolkit"; // NOI18N
-            try {
-                tC = Class.forName(toolkitCN);
-            } catch (ClassNotFoundException ex) {
-                tC = Stage.class.getClassLoader().loadClass(toolkitCN);
-            }
-            g = tC.getMethod("getToolkit"); // NOI18N
-            n = tC.getMethod("enterNestedEventLoop", Object.class); // NOI18N
-            x = tC.getMethod("exitNestedEventLoop", Object.class, Object.class); // NOI18N
-        } catch (ClassNotFoundException | NoSuchMethodException | SecurityException ex) {
-            LOG.log(Level.SEVERE, 
-                "Cannot initialize JavaFX Toolkit access. May cause deadlocks.", 
-                ex
-            );
-        }
-        GET = g;
-        ENTER = n;
-        EXIT = x;
-    }
-
-    private static void enterNestedLoop(HTMLDialogImpl impl) {
-        try {
-            Object tk = GET.invoke(null);
-            ENTER.invoke(tk, impl);
-        } catch (IllegalAccessException | IllegalArgumentException | InvocationTargetException ex) {
-            LOG.log(Level.SEVERE, "enterNestedEventLoop(" + impl + ")", ex);
-        }
-    }
-    
-    private static void exitNestedLoop(HTMLDialogImpl impl) {
-        try {
-            Object tk = GET.invoke(null);
-            EXIT.invoke(tk, impl, null);
-        } catch (IllegalAccessException | IllegalArgumentException | InvocationTargetException ex) {
-            LOG.log(Level.SEVERE, "exitNestedEventLoop(" + impl + ", null)", ex);
-        }
+        return HtmlToolkit.getDefault().convertToComponent(type, pageUrl, loader, onPageLoad, techIds);
     }
 }
