+++ platform/api.htmlui/src/org/netbeans/modules/htmlui/jfx/NbFxPanel.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.netbeans.modules.htmlui;
+package org.netbeans.modules.htmlui.jfx;
 
 import java.awt.im.InputMethodRequests;
 import javafx.embed.swing.JFXPanel;
